+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Display both "download" and "process" permissions (#110)
 -   Download button for models (#110)
 -   Add CSV export (#117)
+-   Compute Plans comparison page (#130)
 
 ### Fixed